Revert "Rename `type_def` to `typedef` for consistency with `opdef`."
This reverts commit 3805b25c86967fe747c8bc1a7fa5913d2e78164d.
