+++ src/resource/type_def.rs
